context: extract service hook
